requirements.txt
python-docx>=1.1.0
olefile>=0.47
pdfplumber>=0.11.0
charset-normalizer>=3.0
openai>=1.0.0
Pillow>=10.0.0
pyyaml>=6.0
rich>=13.0
pytest>=8.0
